fix(form): Fixes
Fixed issues with fields in sections not being sorted according to the arrangement configured. Sounds easy right? No, it wasn't. Here is what has changed:

We now have 2 extra data models that relate an attribute or data element to a program section or program stage section. these models hold the sortOrder for each data item in the section.

So yeah, you are going to need to re sync the metadata.

Fix #17

BREAKING CHANGE: Metadata re-synchronization is required