RefactorD2CoordinateValue o DGeometryValue an addGeoJon suppot
The D2CoordinateValue and related classes have been refactored to D2GeometryValue to better reflect their function. Additionally, a method to generate a GeoJson object from the latitude and longitude values has been added. This change will allow serialized JSON objects to be used in form values labelled 'geometry'.

There's also implementation in form controllers that assigns a geometry to a trackedEntity, enrollment or event if there is a geometry field with key enrollment