feat(data_value): support MULTI_TEXT value type
Added support for the `MULTI_TEXT` value type in `D2DataValue`. When the data element's value type is `MULTI_TEXT`, the `value` will be split by commas, and each part will be looked up in the option set. Then the `displayName` or `name` of found option will be returned. If option is not found, the original `value` will be used instead. Finally the output will be joined by commas.